src/misc/autogen_gis/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
fastapi==0.108.0
FLAML==2.1.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
numpy==1.26.3
openai==1.6.1
packaging==23.2
pyautogen==0.2.3
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
starlette==0.32.0.post1
termcolor==2.4.0
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
